moved board functionality to board crate
